+++ src/main/java/com/blackrook/small/dispatch/DispatchEntryPoint.java
@@ -373,10 +373,8 @@ public class DispatchEntryPoint<S extends DispatchComponent>
 				{
 					if (pinfo.getType() != Cookie.class)
 						throw new ClassCastException("Parameter " + i + " is not a Cookie type.");
-
 					String cookieName = pinfo.getName();
 					Cookie c = cookieMap.containsKey(cookieName) ? cookieMap.get(cookieName) : new Cookie(cookieName, "");
-					response.addCookie(c);
 					invokeParams[i] = c;
 					break;
 				}